requirements.txt
flask
flask-cors
flask-sqlalchemy
flask-migrate
flask-socketio
torch
torchvision
opencv-python
pillow
requests
twilio
python-dotenv
numpy
geopy
gunicorn
ultralytics>=8.0.0
